pyproject.toml
[tool.poetry]
name = "dpa"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.2"
django-phonenumber-field = "^7.1.0"
django-phonenumbers = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
